Conflicts:
	DB/dbx/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbx/definition/mybatis/param/relate/KaigoJigyoshaMapperParameter.java
	DB/dbx/standard/definition/src/main/java/jp/co/ndensan/reams/db/dbx/definition/mybatis/param/relate/KaigoServiceShuruiMapperParameter.java
	DB/dbx/standard/definition/src/test/java/jp/co/ndensan/reams/db/dbx/definition/mybatis/param/relate/KaigoJigyoshaMapperParameterTest.java
	DB/dbx/standard/definition/src/test/java/jp/co/ndensan/reams/db/dbx/definition/mybatis/param/relate/KaigoServiceShuruiMapperParameterTest.java
	DB/dbx/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbx/entity/db/relate/relate/KaigoJigyoshaEntity.java
	DB/dbx/standard/entity/src/main/java/jp/co/ndensan/reams/db/dbx/entity/db/relate/relate/KaigoServiceShuruiEntity.java
	DB/dbx/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbx/persistence/db/mapper/relate/relate/IKaigoJigyoshaMapper.java
	DB/dbx/standard/persistence/src/main/java/jp/co/ndensan/reams/db/dbx/persistence/db/mapper/relate/relate/IKaigoServiceShuruiMapper.java
	DB/dbx/standard/persistence/src/test/java/jp/co/ndensan/reams/db/dbx/persistence/db/mapper/relate/relate/IKaigoJigyoshaMapperTest.java
	DB/dbx/standard/persistence/src/test/java/jp/co/ndensan/reams/db/dbx/persistence/db/mapper/relate/relate/IKaigoServiceShuruiMapperTest.java
	DB/dbx/standard/service/src/main/java/jp/co/ndensan/reams/db/dbx/service/core/relate/KaigoJigyoshaManager.java
	DB/dbx/standard/service/src/main/java/jp/co/ndensan/reams/db/dbx/service/core/relate/KaigoServiceShuruiManager.java
	DB/dbx/standard/service/src/test/java/jp/co/ndensan/reams/db/dbx/service/core/relate/KaigoJigyoshaManagerTest.java
	DB/dbx/standard/service/src/test/java/jp/co/ndensan/reams/db/dbx/service/core/relate/KaigoServiceShuruiManagerTest.java
	DB/pom.xml

+++ DB/DBC/standard/model/src/main/java/jp/co/ndensan/reams/db/dbc/model/KyufuhiKashitsukekinShinseiModel.java
@@ -18,6 +18,7 @@ import jp.co.ndensan.reams.uz.uza.biz.YMDHMS;
 import jp.co.ndensan.reams.uz.uza.biz.YubinNo;
 import jp.co.ndensan.reams.uz.uza.lang.FlexibleDate;
 import jp.co.ndensan.reams.uz.uza.lang.RString;
+import jp.co.ndensan.reams.uz.uza.math.Decimal;
 import jp.co.ndensan.reams.uz.uza.util.db.EntityDataState;
 
 /**
@@ -311,7 +312,7 @@ public class KyufuhiKashitsukekinShinseiModel implements Serializable {
      *
      * @return 提出物管理番号
      */
-    public int get提出物管理番号() {
+    public Decimal get提出物管理番号() {
         return entity.getTeishutsubutsuKanriNo();
     }
 
@@ -590,7 +591,7 @@ public class KyufuhiKashitsukekinShinseiModel implements Serializable {
      *
      * @param 提出物管理番号 提出物管理番号
      */
-    public void set提出物管理番号(int 提出物管理番号) {
+    public void set提出物管理番号(Decimal 提出物管理番号) {
         requireNonNull(提出物管理番号, UrSystemErrorMessages.値がnull.getReplacedMessage("提出物管理番号"));
         entity.setTeishutsubutsuKanriNo(提出物管理番号);
     }
